Refactor: Modular Skeleton States
Refactored the PageSkeleton system into specialized variants (grid, post, simple) to reduce Cumulative Layout Shift (CLS).

Key changes:
- Modularized PageSkeleton.tsx with variant sub-components.
- Introduced RouteConfig interface in a shared types file to prevent circular dependencies.
- Enhanced routes.ts with skeleton metadata.
- Implemented decoupled path-to-skeleton resolution in utils.ts with regex support for parameterized routes.
- Integrated dynamic skeleton selection into RootLayout.
- Updated visual regression snapshots to align with the improved loading states.

Verified with pnpm audit, lint, build, and functional E2E tests.
